# Performance improvement PersonaBar pages treeview on Expand/Collapse